Move metadata stuff to own module
